Removing old conv2d implementation
