+++ src/tensor_ops/conv2d/cpu_kernel.rs
@@ -1,6 +1,5 @@
-use crate::prelude::Tensorlike;
 use crate::shapes::{Dtype, Shape};
-use crate::tensor::{cpu::*, Tensor, ZerosTensor};
+use crate::tensor::{cpu::*, *};
 use crate::tensor_ops::matmul::cpu_kernel::MatMulImpl;
 
 use super::{Conv2DKernel, Conv2DOp};
@@ -11,10 +10,10 @@ impl Conv2DOp {
     #[inline(always)]
     fn unfold_idx(&self, [k1, k2, y, x]: [usize; 4]) -> Option<[usize; 2]> {
         let mut oh = y + self.padding;
-        if oh < k1 {
+        if oh < self.dilation * k1 {
             return None;
         }
-        oh -= k1;
+        oh -= self.dilation * k1;
         if oh % self.stride != 0 {
             return None;
         }
@@ -24,10 +23,10 @@ impl Conv2DOp {
         }
 
         let mut ow = x + self.padding;
-        if ow < k2 {
+        if ow < self.dilation * k2 {
             return None;
         }
-        ow -= k2;
+        ow -= self.dilation * k2;
         if ow % self.stride != 0 {
             return None;
         }
@@ -42,7 +41,7 @@ impl Conv2DOp {
 
 impl Cpu {
     #[inline]
-    fn conv2d_forward<E: Dtype>(
+    fn fwd<E: Dtype>(
         &self,
         op: &Conv2DOp,
         img: &[E],
@@ -55,19 +54,16 @@ impl Cpu {
     {
         {
             let mut i = 0;
-            for c in 0..op.chan_in {
+            for c in 0..(op.groups * op.chan_in) {
                 for k1 in 0..op.kernel {
                     for k2 in 0..op.kernel {
                         for oh in 0..op.h_out {
-                            let y = (oh * op.stride + k1).wrapping_sub(op.padding);
+                            let y = (oh * op.stride + op.dilation * k1).wrapping_sub(op.padding);
                             for ow in 0..op.w_out {
-                                let x = (ow * op.stride + k2).wrapping_sub(op.padding);
+                                let x =
+                                    (ow * op.stride + op.dilation * k2).wrapping_sub(op.padding);
                                 if y < op.h_in && x < op.w_in {
-                                    unsafe {
-                                        *buf.get_unchecked_mut(i) = *img.get_unchecked(
-                                            c * (op.w_in * op.h_in) + y * op.w_in + x,
-                                        )
-                                    };
+                                    buf[i] = img[c * (op.w_in * op.h_in) + y * op.w_in + x];
                                 }
                                 i += 1;
                             }
@@ -77,25 +73,27 @@ impl Cpu {
             }
         }
 
-        // (O, C * K * K) * (C * K * K, OH * OW) = (O, OH * OW)
-        let m = op.chan_out;
+        // (G, O / G, C * K * K) * (G, C * K * K, OH * OW) = (G, O / G, OH * OW)
+        let m = op.chan_out / op.groups;
         let k = op.chan_in * op.kernel * op.kernel;
         let n = op.w_out * op.h_out;
-        Self::matmul(
-            (m, k, n),
-            filters.as_ptr(),
-            [k, 1],
-            buf.as_ptr(),
-            [n, 1],
-            out.as_mut_ptr(),
-            [n, 1],
-        );
+        for g in 0..op.groups {
+            Self::matmul(
+                (m, k, n),
+                filters[g * m * k..].as_ptr(),
+                [k, 1],
+                buf[g * k * n..].as_ptr(),
+                [n, 1],
+                out[g * m * n..].as_mut_ptr(),
+                [n, 1],
+            );
+        }
         Ok(())
     }
 
     #[inline]
     #[allow(clippy::too_many_arguments)]
-    fn conv2d_backward<E: Dtype>(
+    fn bwd<E: Dtype>(
         &self,
         op: &Conv2DOp,
         img: &[E],
@@ -116,11 +114,8 @@ impl Cpu {
                         for y in 0..op.h_in {
                             for x in 0..op.w_in {
                                 if let Some([oh, ow]) = op.unfold_idx([k1, k2, y, x]) {
-                                    unsafe {
-                                        *buf.get_unchecked_mut(i) = *grad_out.get_unchecked(
-                                            o * (op.h_out * op.w_out) + oh * op.w_out + ow,
-                                        );
-                                    }
+                                    buf[i] =
+                                        grad_out[o * (op.h_out * op.w_out) + oh * op.w_out + ow];
                                 }
                                 i += 1;
                             }
@@ -132,36 +127,40 @@ impl Cpu {
 
         {
             // img_g += filters^T * unfold(grad_out)
-            // (C, H * W) += (C, O * K * K) * (O * K * K, H * W)
+            // (G, C, H * W) += (G, C, O/G * K * K) * (G, O/G * K * K, H * W)
             let m = op.chan_in;
-            let k = op.chan_out * op.kernel * op.kernel;
+            let k = (op.chan_out / op.groups) * op.kernel * op.kernel;
             let n = op.h_in * op.w_in;
-            Self::matmul(
-                (m, k, n),
-                filters_tr.as_ptr(),
-                [k, 1],
-                buf.as_ptr(),
-                [n, 1],
-                grad_img.as_mut_ptr(),
-                [n, 1],
-            );
+            for g in 0..op.groups {
+                Self::matmul(
+                    (m, k, n),
+                    filters_tr[g * m * k..].as_ptr(),
+                    [k, 1],
+                    buf[g * k * n..].as_ptr(),
+                    [n, 1],
+                    grad_img[g * m * n..].as_mut_ptr(),
+                    [n, 1],
+                );
+            }
         }
 
         {
-            // weight_g^T += img * patches^T
-            // (C, O * K * K) += (C, H * W) * (H * W, O * K * K)
+            // weight_g^T += img * unfold(patches)^T
+            // (G, C, O/G * K * K) += (G, C, H * W) * (G, H * W, O/G * K * K)
             let m = op.chan_in;
             let k = op.h_in * op.w_in;
-            let n = op.chan_out * op.kernel * op.kernel;
-            Self::matmul(
-                (m, k, n),
-                img.as_ptr(),
-                [k, 1],
-                buf.as_ptr(),
-                [1, k],
-                grad_filters_tr.as_mut_ptr(),
-                [n, 1],
-            );
+            let n = (op.chan_out / op.groups) * op.kernel * op.kernel;
+            for g in 0..op.groups {
+                Self::matmul(
+                    (m, k, n),
+                    img[g * m * k..].as_ptr(),
+                    [k, 1],
+                    buf[g * k * n..].as_ptr(),
+                    [1, k],
+                    grad_filters_tr[g * m * n..].as_mut_ptr(),
+                    [n, 1],
+                );
+            }
         }
         Ok(())
     }
@@ -182,7 +181,14 @@ where
         rhs: &Tensor<R, E, Self>,
         out: &mut Tensor<O, E, Self>,
     ) -> Result<(), Self::Err> {
-        let mut patches = self.try_alloc_zeros::<E>(op.inp_patches_shape().num_elements())?;
+        let patches = (
+            op.groups * op.chan_in,
+            op.kernel,
+            op.kernel,
+            op.h_out,
+            op.w_out,
+        );
+        let mut patches = self.try_alloc_zeros::<E>(patches.num_elements())?;
         let [lstride, ostride] = match L::NUM_DIMS {
             3 => [0; 2],
             4 => [lhs.strides[0], out.strides[0]],
@@ -192,7 +198,7 @@ where
         let rhs = rhs.data.as_ref();
         let out = Arc::make_mut(&mut out.data);
         for i_batch in 0..op.batch {
-            self.conv2d_forward(
+            self.fwd(
                 &op,
                 &lhs[i_batch * lstride..],
                 rhs,
@@ -213,8 +219,15 @@ where
         out: &impl Tensorlike<O, E, Self>,
         grad_out: &Self::Vec<E>,
     ) -> Result<(), Self::Err> {
-        let f_tr_shape = op.filters_tr_shape();
-        let mut patches = self.try_alloc_zeros::<E>(op.out_patches_shape().num_elements())?;
+        let f_tr_shape = [
+            op.groups,
+            op.chan_in,
+            op.chan_out / op.groups,
+            op.kernel,
+            op.kernel,
+        ];
+        let patches_shape = [op.chan_out, op.kernel, op.kernel, op.h_in, op.w_in];
+        let mut patches = self.try_alloc_zeros::<E>(patches_shape.num_elements())?;
         let mut f1023 = self.try_alloc_zeros::<E>(f_tr_shape.num_elements())?;
         let mut grad_f1023 = self.try_alloc_zeros::<E>(f_tr_shape.num_elements())?;
 
@@ -222,8 +235,8 @@ where
             // transpose filters in f1023
             let buf = rhs.data.as_ref();
             let mut f_idx = NdIndex::new(f_tr_shape, f_tr_shape.strides());
-            while let Some((i, [c, o, k1, k2])) = f_idx.next_with_idx() {
-                let idx = o * rhs.strides[0]
+            while let Some((i, [g, c, o, k1, k2])) = f_idx.next_with_idx() {
+                let idx = (g * (op.chan_out / op.groups) + o) * rhs.strides[0]
                     + c * rhs.strides[1]
                     + k1 * rhs.strides[2]
                     + k2 * rhs.strides[3];
@@ -239,7 +252,7 @@ where
         let lhs = lhs.data.as_ref();
 
         for i_batch in 0..op.batch {
-            self.conv2d_backward(
+            self.bwd(
                 &op,
                 &lhs[i_batch * lstride..],
                 &mut grad_lhs[i_batch * lstride..],
@@ -253,8 +266,8 @@ where
         {
             // untranspose filters
             let mut f_idx = NdIndex::new(f_tr_shape, f_tr_shape.strides());
-            while let Some((i, [c, o, k1, k2])) = f_idx.next_with_idx() {
-                let idx = o * rhs.strides[0]
+            while let Some((i, [g, c, o, k1, k2])) = f_idx.next_with_idx() {
+                let idx = (g * (op.chan_out / op.groups) + o) * rhs.strides[0]
                     + c * rhs.strides[1]
                     + k1 * rhs.strides[2]
                     + k2 * rhs.strides[3];
